requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
pytest>=8.0
pytest-django>=4.8